PR feedback. start fixing issues.
