+++ src/libraries/Microsoft.PowerFx.Core/Public/Types/FormulaType.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using Microsoft.PowerFx.Core.Types;
 using Microsoft.PowerFx.Core.Utils;
@@ -15,7 +14,7 @@ namespace Microsoft.PowerFx.Core.Public.Types
     /// Formula Types are a class hiearchy.
     /// </summary>
     [DebuggerDisplay("{_type}")]
-    [ImmutableObject(true)]
+    [ThreadSafeImmutable]
     public abstract class FormulaType
     {
         // protected isn't enough to let derived classes access this.
