+++ src/libraries/Microsoft.PowerFx.Core/Texl/Intellisense/AddSuggestionHelper.cs
@@ -8,6 +8,7 @@ using Microsoft.PowerFx.Core.Utils;
 
 namespace Microsoft.PowerFx.Core.Texl.Intellisense
 {
+    [ThreadSafeImmutable]
     internal class AddSuggestionHelper
     {
         public bool AddSuggestion(IntellisenseData.IntellisenseData intellisenseData, string suggestion, SuggestionKind suggestionKind, SuggestionIconKind iconKind, DType type, bool requiresSuggestionEscaping, uint sortPriority = 0)
