+++ src/tests/Microsoft.PowerFx.Core.Tests/ThreadingTests.cs
@@ -1,0 +1,71 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using Xunit;
+
+namespace Microsoft.PowerFx.Core.Tests
+{
+    // Do static analysis to look for potential threading issues. 
+    public class ThreadingTests
+    {
+        [Fact]
+        public void CheckFxCore()
+        {
+            var asm = typeof(Core.Syntax.Nodes.TexlNode).Assembly;
+
+            var bugsFieldType = new HashSet<Type>
+            {            
+                // make these fields readonly
+                typeof(Core.Localization.TexlStrings.StringGetter),
+                typeof(Core.Localization.ErrorResourceKey),
+            };
+
+            var bugNames = new HashSet<string>
+            {
+                // readonly arrays / dictionary - is there an IReadOnly type to changes these too instead? 
+                "LazyList`1.Empty",
+                "DType._kindToSuperkindMapping",
+                "DTypeSpecParser._types",
+                "BuiltinFunctionsCore._library",
+                "ArgumentSuggestions._languageCodeSuggestions",
+                "IntellisenseProvider.SuggestionHandlers",
+                "DateAddFunction.SubDayStringList",
+                "PrettyPrintVisitor.BinaryPrecedence",
+                "ErrorResource.ErrorResourceTagToReswSuffix",
+                "StringResources.Strings",
+                "StringResources.ErrorResources",
+                "DelegationCapability._binaryOpToDelegationCapabilityMap",
+                "DelegationCapability._unaryOpToDelegationCapabilityMap",
+                "DelegationCapability._operatorToDelegationCapabilityMap",
+                "ODataFunctionMappings.BinaryOpToOperatorMap",
+                "ODataFunctionMappings.UnaryOpToOperatorMap",
+                       
+                // Potential bugs, Need more review...
+                "ArgumentSuggestions.CustomFunctionSuggestionProviders",
+                "TrackingProvider.Instance",
+                "TexlLexer._prebuiltLexers",
+
+                "DataTypeInfo.NoValidFormat", // returns arrays 
+                "DataTypeInfo.AllowedValuesOnly",
+                "DataTypeInfo._validDataFormatsPerDKind",
+
+                "FeatureFlags.StringInterpolation",
+                "EmptyEnumerator`1._instance",
+                "StringBuilderCache`1.maxBuilderSize",
+                "StringBuilderCache`1.cachedInstance",
+                "Contracts._assertFailExCtor",
+                "CurrentLocaleInfo.<CurrentLocaleName>k__BackingField",
+                "CurrentLocaleInfo.<CurrentUILanguageName>k__BackingField",
+                "CurrentLocaleInfo.<CurrentLCID>k__BackingField",
+                "StringResources.<ExternalStringResources>k__BackingField",
+                "StringResources.<ShouldThrowIfMissing>k__BackingField",
+                "TexlLexer._lex",
+                "DelegationCapability.maxSingleCapabilityValue",
+            };
+
+            AnalyzeThreadSafety.CheckStatics(asm, bugsFieldType, bugNames);
+        }
+    }
+}
