+++ src/tests/Microsoft.PowerFx.Interpreter.Tests/ThreadingTests.cs
@@ -1,0 +1,29 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using Microsoft.PowerFx.Core.Tests;
+using Xunit;
+
+namespace Microsoft.PowerFx.Interpreter.Tests
+{
+    // Do static analysis to look for potential threading issues. 
+    public class ThreadingTests
+    {
+        [Fact]
+        public void CheckIntepreter()
+        {
+            var asm = typeof(RecalcEngine).Assembly;
+            var bugsFieldType = new HashSet<Type>
+            {                
+            };
+
+            var bugNames = new HashSet<string>
+            {
+            };
+
+            AnalyzeThreadSafety.CheckStatics(asm, bugsFieldType, bugNames);
+        }
+    }
+}
